Optimizing RGF
