+++ src/3_matrices/recursive-greens/Module.jl
@@ -1,6 +1,7 @@
 module RecursiveGreensModule
 
+import Base: +, -
 include("RGF.jl") 
-export CreateBlockMatrix, BlockMatrix, getInvRGF, getInvJulia, getInvRGFDiagonal
+export CreateBlockMatrix, ToBlockMatrix, BlockMatrix, getInvRGF!, getInvJulia!, getInvRGFDiagonal!, +, -
 
 end
