+++ src/3_matrices/recursive-greens/RGF.jl
@@ -3,8 +3,18 @@ using LinearAlgebra
 
 include("BlockMatrix.jl")
 
+function getInvJulia!(matrixObject::BlockMatrix)
+    matrixObject.matrix = inv(Matrix(matrixObject.matrix))
+end
+
 # Function to compute the inverse of a block matrix using the recursive Green's function (RGF) method.
-function getInvRGF(matrixObject::BlockMatrix)
+function getInvRGF!(matrixObject::BlockMatrix)
+    # does not work if there is only one block
+    if matrixObject.numBlocks == 1
+        getInvJulia!(matrixObject)
+        return
+    end
+
     # Compute the forward and backward generators.
     forwardGen, backwardGen = computeGenerators(matrixObject)
 
@@ -15,15 +25,19 @@ function getInvRGF(matrixObject::BlockMatrix)
     computeBottomBlocks(matrixObject, backwardGen, lastDiag, sparseBuild)
     # display(sparse(sparseBuild.rowIndices, sparseBuild.columnIndices, sparseBuild.values, matrixObject.matrixSize, matrixObject.matrixSize))
     
-    matrix = deepcopy(matrixObject)
-    matrix.matrix = sparse(sparseBuild.rowIndices, sparseBuild.columnIndices, sparseBuild.values, matrixObject.matrixSize, matrixObject.matrixSize)
+    matrixObject.matrix = sparse(sparseBuild.rowIndices, sparseBuild.columnIndices, sparseBuild.values, matrixObject.matrixSize, matrixObject.matrixSize)
     # Return the diagonal blocks along with the top and bottom row blocks of the inverse matrix.
     # return diag, computeTopBlocks(matrixObject, forwardGen, diag[1]), computeBottomBlocks(matrixObject, backwardGen, diag[end])
-    return matrix
 end
 
 # Function to perform the getInvRGF function except without the top and bottom rows of the matrix
-function getInvRGFDiagonal(matrixObject::BlockMatrix)
+function getInvRGFDiagonal!(matrixObject::BlockMatrix)
+    # does not work if there is only one block
+    if matrixObject.numBlocks == 1
+        getInvJulia!(matrixObject)
+        return
+    end
+
     # Compute the forward and backward generators.
     _, backwardGen = computeGenerators(matrixObject)
 
@@ -31,13 +45,5 @@ function getInvRGFDiagonal(matrixObject::BlockMatrix)
     # Compute the diagonal blocks of the inverse matrix.
     computeDiagonalBlocks(matrixObject, backwardGen, sparseBuild)
     
-    matrix = deepcopy(matrixObject)
-    matrix.matrix = sparse(sparseBuild.rowIndices, sparseBuild.columnIndices, sparseBuild.values, matrixObject.matrixSize, matrixObject.matrixSize)
-    return matrix
-end
-
-function getInvJulia(matrixObject::BlockMatrix)
-    matrix = deepcopy(matrixObject)
-    matrix.matrix = inv(Matrix(matrixObject.matrix))
-    return matrix
+    matrixObject.matrix = sparse(sparseBuild.rowIndices, sparseBuild.columnIndices, sparseBuild.values, matrixObject.matrixSize, matrixObject.matrixSize)
 end
